+++ docs/changelog.md
@@ -15,4 +15,5 @@
   - added `T_COMMA` token
   - added `T_DOUBLE_AT` token
   - added `T_EMPTY` token
+  - added `T_TILDE` lazy-match token
   - added `T_UNDERSCORE` lazy-match token